+++ src/Psl/Exception/UnderflowException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Psl\Exception;
+
+use UnderflowException as UnderflowRootException;
+
+class UnderflowException extends UnderflowRootException implements ExceptionInterface
+{
+}